testing makefile (namechanging function) done. [need to study about Mock], [for test, must Uncomment files.service.spec.ts]
