--- src/helpers/image-storage.ts
@@ -1,14 +1,0 @@
-import { diskStorage } from 'multer';
-import path = require('path');
-import { v4 as uuid } from 'uuid';
-
-export const saveImageToStorage = {
-  storage: diskStorage({
-    destination: './storage/uploads',
-    filename: (req, file, cb) => {
-      const fileExtension: string = path.extname(file.originalname);
-      const fileName: string = uuid() + fileExtension;
-      cb(null, fileName);
-    },
-  }),
-};
